Koda/requirements.txt
numpy==1.26.4
python-docx==1.1.2
